+++ test/validator/samples/animation-each-with-whitespace/input.svelte
@@ -1,0 +1,7 @@
+<script>
+	function flip() {}
+</script>
+
+<div>
+	{#each [] as n (n)} <div animate:flip /> {/each}
+</div>